READ DESCRIP. PLEASE
I commented most of the scripts

I put a '//NEEDS COMMENTING' at the top of scripts that still needs to be commented (eg. Dhan's classes & Lachlan's UI & Audio stuff) which y'all need to do yourself.

What i dd for most of the scripts, was seperate all public and private variables for neatness, and put them in a region titled 'Variables'. I wrote who significantly contributed to that script, and if there was more than one person i put their contribution in paranthesis. And finally, i wrote a comment above each function for what it does, and collapsed everything for neatness again lmao.
I tried to steer clear of comments inside functions.

Refer to the AI scripts if you didn't understand, lol.

If y'all can follow this structure for consistency, that would be appreciated. Except for dhans procedural scripts, that's on a whole other level.

And if you find any scripts that you are 100% sure are not needed (Eg, the quarter million winGame scripts we have for some reason LMAO), please delete them, and if any scripts are not in the right folder, please relocate them.

Thanks for tuning into my TED talk.